+++ changelog.md
@@ -1,8 +1,9 @@
 - REQUIRES [CraterLib](https://www.curseforge.com/minecraft/mc-mods/craterlib) - [Modrinth](https://modrinth.com/mod/craterlib)
-- [Migration guide](https://sdlink.fdd-docs.com/migration/) for V2 users
 - [Online Config Editor](https://editor.firstdark.dev)
+- [Documentation](https://sdlink.fdd-docs.com)
+- This single jar works on 1.18.2-1.21.4
 
-*Requires CraterLib 2.1.2 or newer*
+*Requires CraterLib 2.1.3 or newer*
 
 **Bug Fixes**:
 
@@ -16,6 +17,9 @@
 **New Features**:
 
 - Implemented a basic message spam detector
+- Added Optional verification, to allow people to use most (not all) features that required access control to be enabled, without it being enabled.
+- Added basic spam checker to prevent spammed messages from being relayed to discord. Should help with command spams and ratelimits
+- Allow Verify and Unverify commands to be used in DM, and verification codes can also be DM'ed to the bot
 
 **Technical Changes**: